pyproject.toml
[tool.poetry]
name = "geo-processor"
version = "1.0.0"
description = "Compute centroid and bounding box for a list of coordinates."
authors = ["Tu Nombre <tu@email.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "0.115.0"
uvicorn = {extras = ["standard"], version = "0.30.6"}
pydantic = "2.9.2"
python-decouple = "3.8"
pydantic-settings = "2.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
httpx = "0.27.2"
ruff = "0.6.9"
mypy = "1.11.2"
pre-commit = "3.8.0"
types-requests = "*"

[tool.poetry.scripts]
geo-processor = "app.main:app"

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
target-version = "py310"
line-length = 100
fix = true

[tool.ruff.lint]
select = ["F","E","W","I","UP","B","C4","SIM","PL","RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
check_untyped_defs = true
exclude = ["^tests/.*"]
plugins = []

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
